extracted message header buttons into a separate file
